Added measure_distance script, updated readme
